move windows compilmations to an independant package
