+++ keopswindows/src/keopswindows/compile_nvrtc_jit.py
@@ -9,10 +9,10 @@ from .detection import (
 from .utils import find_package_location
 
 
-def compile_pykeops_nvrtc(build_folder):
+def compile_nvrtc_jit(build_folder):
 
-    pykeops_dir = Path(find_package_location("pykeops")).parent
-    source_file = pykeops_dir / "common" / "keops_io" / "pykeops_nvrtc.cpp"
+    keops_dir = Path(find_package_location("keopscore")).parent
+    source_file = keops_dir / "binders" / "nvrtc" / "nvrtc_jit.cpp"
 
     macros = [
         "-DMAXIDGPU=0",
@@ -26,11 +26,12 @@ def compile_pykeops_nvrtc(build_folder):
     compile(
         source_file=source_file,
         macros=macros,
-        includes=[include_dirs[key] for key in ["python", "pybind11", "keops", "cuda"]],
-        link_dirs=[lib_dirs[key] for key in ["python", "cuda"]],
-        links=[lib_names[key] for key in ["cuda", "nvrtc", "cudart", "python"]],
-        suffix=".pyd",
+        includes=[include_dirs[key] for key in ["keops", "cuda"]],
+        link_dirs=[lib_dirs[key] for key in ["cuda"]],
+        links=[lib_names[key] for key in ["cuda", "nvrtc", "cudart"]],
+        suffix=".dll",
         output_dir=build_folder,
         print_cmakelists=False,
         show_cmake_commands_output=False,
+        clean_tmp_build_dir=False,
     )
