+++ keopswindows/src/keopswindows/detection.py
