+++ keopswindows/src/keopswindows/templates/CMakeLists.txt
